refactor(pastas): organiza estrutura dos arquivos
