+++ .env.example
@@ -1,0 +1,2 @@
+DB_USER=root
+DB_PASSWORD=admin
