+++ src/main/java/app/models/EquipeModel.java
@@ -1,6 +1,5 @@
-package app.controllers;
+package app.models;
 
-import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
 
 public class EquipeModel {
