+++ src/main/java/module-info.java
@@ -7,4 +7,8 @@ module app.controllers {
     opens student to javafx.fxml;
     opens app.controllers to javafx.fxml;
     exports app.controllers;
+    exports app.models;
+    opens app.models to javafx.fxml;
+    exports app.helpers;
+    opens app.helpers to javafx.fxml;
 }
